+++ packages/graph/src/index.ts
@@ -11,4 +11,6 @@ export * from './themes';
 export * from './tooltip/Tooltip';
 export * from './tooltip/TooltipItem';
 export * from './utils/getAxisScale';
+export * from './utils/getColorScale';
+export * from './utils/getDataGroupByEncodings';
 export * from './utils/getRecordFieldSelector';